+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+v2.0.1
+------
+
+- Remove cyclic reference involving RequestCycle and thus the socket resource, which led to the socket only being freed after cycle collector kicked in.
+
 v2.0.0
 ------